+++ ptai-jenkins-plugin/src/main/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/plugin/jenkins/JenkinsAstJob.java
@@ -1,6 +1,8 @@
 package com.ptsecurity.appsec.ai.ee.utils.ci.integration.plugin.jenkins;
 
+import com.ptsecurity.appsec.ai.ee.utils.ci.integration.exceptions.AstPolicyViolationException;
 import com.ptsecurity.appsec.ai.ee.utils.ci.integration.exceptions.GenericException;
+import com.ptsecurity.appsec.ai.ee.utils.ci.integration.exceptions.MinorAstErrorsException;
 import com.ptsecurity.appsec.ai.ee.utils.ci.integration.functions.TextOutput;
 import com.ptsecurity.appsec.ai.ee.utils.ci.integration.jobs.GenericAstJob;
 import com.ptsecurity.appsec.ai.ee.utils.ci.integration.operations.JsonAstJobSetupOperationsImpl;
@@ -8,6 +10,7 @@ import com.ptsecurity.appsec.ai.ee.utils.ci.integration.operations.UiAstJobSetup
 import com.ptsecurity.appsec.ai.ee.utils.ci.integration.plugin.jenkins.operations.JenkinsAstOperations;
 import com.ptsecurity.appsec.ai.ee.utils.ci.integration.plugin.jenkins.operations.JenkinsFileOperations;
 import com.ptsecurity.appsec.ai.ee.utils.ci.integration.plugin.jenkins.utils.BuildInfo;
+import com.ptsecurity.appsec.ai.ee.utils.ci.integration.plugin.jenkins.workmode.WorkModeSync;
 import hudson.FilePath;
 import hudson.Launcher;
 import hudson.model.Run;
@@ -21,12 +24,17 @@ import lombok.extern.slf4j.Slf4j;
 
 import java.util.List;
 
+import static com.ptsecurity.appsec.ai.ee.utils.ci.integration.plugin.jenkins.workmode.WorkModeSync.OnAstError.NONE;
+
 @Slf4j
 @Getter
 @Setter
 @SuperBuilder
 @ToString(callSuper = true)
 public class JenkinsAstJob extends GenericAstJob implements TextOutput {
+    @NonNull
+    protected Plugin plugin;
+
     @NonNull
     protected Run<?, ?> run;
     /**
@@ -105,5 +113,12 @@ public class JenkinsAstJob extends GenericAstJob implements TextOutput {
             setupOps = UiAstJobSetupOperationsImpl.builder()
                     .owner(this)
                     .build();
+    }
+
+    @Override
+    public JobExecutionResult processException(@NonNull final GenericException e) {
+        JobExecutionResult res = super.processException(e);
+        plugin.setBuildResult(run, res, e);
+        return res;
     }
 }