+++ ptai-jenkins-plugin/src/main/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/plugin/jenkins/workmode/WorkModeSync.java
@@ -1,26 +1,34 @@
 package com.ptsecurity.appsec.ai.ee.utils.ci.integration.plugin.jenkins.workmode;
 
 import com.ptsecurity.appsec.ai.ee.utils.ci.integration.Resources;
+import com.ptsecurity.appsec.ai.ee.utils.ci.integration.domain.Reports;
+import com.ptsecurity.appsec.ai.ee.utils.ci.integration.jobs.GenericAstJob;
 import com.ptsecurity.appsec.ai.ee.utils.ci.integration.plugin.jenkins.reports.BaseReport;
 import hudson.Extension;
+import hudson.util.ListBoxModel;
 import lombok.Getter;
 import lombok.NonNull;
 import lombok.ToString;
 import org.jenkinsci.Symbol;
+import org.jvnet.localizer.LocaleProvider;
 import org.kohsuke.stapler.DataBoundConstructor;
 
 import java.util.ArrayList;
+import java.util.Locale;
 
 @ToString
 public class WorkModeSync extends WorkMode {
+    public enum OnAstError {
+        NONE, FAIL, UNSTABLE
+    }
     @Extension
     public static final WorkModeDescriptor DESCRIPTOR = new Descriptor();
 
     @Getter
-    private final boolean failIfFailed;
+    private final OnAstError onAstFailed;
 
     @Getter
-    private final boolean failIfUnstable;
+    private final OnAstError onAstUnstable;
 
     @Getter
     private ArrayList<BaseReport> reports;
@@ -34,9 +42,11 @@ public class WorkModeSync extends WorkMode {
 
     @DataBoundConstructor
     public WorkModeSync(
-            final boolean failIfFailed, final boolean failIfUnstable, final ArrayList<BaseReport> reports) {
-        this.failIfFailed = failIfFailed;
-        this.failIfUnstable = failIfUnstable;
+            @NonNull final OnAstError onAstFailed,
+            @NonNull final OnAstError onAstUnstable,
+            final ArrayList<BaseReport> reports) {
+        this.onAstFailed = onAstFailed;
+        this.onAstUnstable = onAstUnstable;
         setReports(reports);
     }
 
@@ -46,6 +56,30 @@ public class WorkModeSync extends WorkMode {
         @Override
         public String getDisplayName() {
             return Resources.i18n_ast_settings_mode_synchronous_label();
+        }
+
+        public ListBoxModel doFillOnAstFailedItems() {
+            ListBoxModel model = new ListBoxModel();
+            model.add(Resources.i18n_ast_settings_mode_synchronous_onastfailed_none(), OnAstError.NONE.name());
+            model.add(Resources.i18n_ast_settings_mode_synchronous_onastfailed_fail(), OnAstError.FAIL.name());
+            model.add(Resources.i18n_ast_settings_mode_synchronous_onastfailed_unstable(), OnAstError.UNSTABLE.name());
+            return model;
+        }
+
+        public ListBoxModel doFillOnAstUnstableItems() {
+            ListBoxModel model = new ListBoxModel();
+            model.add(Resources.i18n_ast_settings_mode_synchronous_onastunstable_none(), OnAstError.NONE.name());
+            model.add(Resources.i18n_ast_settings_mode_synchronous_onastunstable_fail(), OnAstError.FAIL.name());
+            model.add(Resources.i18n_ast_settings_mode_synchronous_onastunstable_unstable(), OnAstError.UNSTABLE.name());
+            return model;
+        }
+
+        public static OnAstError getDefaultOnAstFailed() {
+            return OnAstError.FAIL;
+        }
+
+        public static OnAstError getDefaultOnAstUnstable() {
+            return OnAstError.NONE;
         }
     }