+++ ptai-jenkins-plugin/src/main/resources/com/ptsecurity/appsec/ai/ee/utils/ci/integration/plugin/jenkins/reports/RawData/config.groovy
@@ -1,4 +1,4 @@
-package com.ptsecurity.appsec.ai.ee.utils.ci.integration.plugin.jenkins.reports.Report
+package com.ptsecurity.appsec.ai.ee.utils.ci.integration.plugin.jenkins.reports.RawData
 
 import lib.FormTagLib